+++ brain/systems/cycles/service.py
@@ -242,10 +242,8 @@ async def _async_attach_open_ask_stragglers(
         or launch_context.get("run_kind") != SCHEDULED_DIGEST_RUN_KIND
     ):
         return []
-    from brain.systems.runs.open_ask_digest import (
-        expire_stale_run_deferrals,
-        list_open_ask_stragglers,
-    )
+    from brain.systems.runs.open_ask_digest import list_open_ask_stragglers
+    from brain.systems.runs.run_deferrals import expire_stale_run_deferrals
 
     try:
         async with session.begin_nested():
