+++ brain/systems/runs/open_ask_digest.py
@@ -1,40 +1,18 @@
-"""Expiry policy and coordinator-digest projection for open asks."""
+"""Read-only coordinator-digest projection for open asks."""
 
 from __future__ import annotations
 
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timedelta
 from typing import Any
 
 from sqlalchemy import and_, func, or_, select
 
-from brain.contracts.statuses import (
-    ACTIVE_OPEN_ASK_STATUS_VALUES,
-    AGENT_RUN_DB_STATUS_VALUES,
-    TERMINAL_RUN_STATUS_VALUES,
-    OpenAskStatus,
-    project_run_status_value,
-)
-from brain.platform.db.models.agent_run import AgentRunRow
+from brain.contracts.statuses import ACTIVE_OPEN_ASK_STATUS_VALUES, OpenAskStatus
+from brain.kernel.common.time import assume_utc
 from brain.platform.db.models.open_ask import ObligationKind, OpenAsk
-from brain.systems.runs.open_ask_settlement import _supersede_pending_notices
 
 
 OPEN_ASK_STRAGGLER_AFTER = timedelta(hours=1)
-# Three days is well below the observed 146-176h failures while allowing recovery.
-RUN_DEFERRAL_EXPIRY_AFTER = timedelta(hours=72)
-_TERMINAL_ORIGIN_RUN_STATUS_VALUES = tuple(
-    status
-    for status in AGENT_RUN_DB_STATUS_VALUES
-    if project_run_status_value(status) in TERMINAL_RUN_STATUS_VALUES
-)
-
-
-def _aware_utc(value: datetime | None) -> datetime:
-    if value is None:
-        return datetime.now(timezone.utc)
-    if value.tzinfo is None:
-        return value.replace(tzinfo=timezone.utc)
-    return value.astimezone(timezone.utc)
 
 
 def _age_label(age: timedelta) -> str:
@@ -49,51 +27,8 @@ def _age_label(age: timedelta) -> str:
 
 def _open_ask_age_started_at(row: OpenAsk) -> datetime:
     if row.status == OpenAskStatus.ROUTED.value and row.routed_at is not None:
-        return _aware_utc(row.routed_at)
-    return _aware_utc(row.opened_at)
-
-
-async def expire_stale_run_deferrals(
-    session: Any,
-    *,
-    org_id: str,
-    now: datetime | None = None,
-    expiry_after: timedelta = RUN_DEFERRAL_EXPIRY_AFTER,
-) -> int:
-    """Expire old run promises only after their originating run is terminal."""
-
-    current = _aware_utc(now)
-    cutoff = current - expiry_after
-    results = list(
-        (
-            await session.execute(
-                select(OpenAsk, AgentRunRow.status)
-                .join(AgentRunRow, AgentRunRow.id == OpenAsk.origin_run_id)
-                .where(
-                    OpenAsk.org_id == str(org_id),
-                    OpenAsk.obligation_kind == ObligationKind.RUN_DEFERRAL,
-                    OpenAsk.status == OpenAskStatus.OPEN.value,
-                    OpenAsk.opened_at < cutoff,
-                    AgentRunRow.status.in_(_TERMINAL_ORIGIN_RUN_STATUS_VALUES),
-                )
-                .order_by(OpenAsk.id.asc())
-                .with_for_update(of=OpenAsk)
-            )
-        ).all()
-    )
-    expiry_hours = int(expiry_after.total_seconds() // 3600)
-    for row, run_status in results:
-        row.status = OpenAskStatus.EXPIRED.value
-        row.expired_at = current
-        row.status_reason = (
-            f"Origin run {int(row.origin_run_id)} is terminal ({str(run_status)}); "
-            f"run deferral expired after {expiry_hours}h."
-        )
-    rows = [row for row, _run_status in results]
-    await _supersede_pending_notices(session, [int(row.id) for row in rows])
-    if rows:
-        await session.flush()
-    return len(rows)
+        return assume_utc(row.routed_at)
+    return assume_utc(row.opened_at)
 
 
 async def list_open_ask_stragglers(
@@ -103,7 +38,7 @@ async def list_open_ask_stragglers(
     now: datetime | None = None,
     older_than: timedelta = OPEN_ASK_STRAGGLER_AFTER,
 ) -> list[dict[str, Any]]:
-    current = _aware_utc(now)
+    current = assume_utc(now)
     cutoff = current - older_than
     rows = list(
         (
@@ -161,7 +96,5 @@ async def list_open_ask_stragglers(
 
 __all__ = [
     "OPEN_ASK_STRAGGLER_AFTER",
-    "RUN_DEFERRAL_EXPIRY_AFTER",
-    "expire_stale_run_deferrals",
     "list_open_ask_stragglers",
 ]
